+++ config/test.exs
@@ -1,6 +1,7 @@
 use Mix.Config
 
 alias Commanded.EventStore.Adapters.InMemory
+alias Commanded.Serialization.JsonSerializer
 
 config :logger, :console, level: :warn, format: "[$level] $message\n"
 
@@ -12,5 +13,5 @@ config :commanded,
   dispatch_consistency_timeout: 100,
   event_store_adapter: InMemory,
   reset_storage: fn ->
-    {:ok, _event_store} = InMemory.start_link(serializer: Commanded.Serialization.JsonSerializer)
+    {:ok, _event_store} = InMemory.start_link(serializer: JsonSerializer)
   end