+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Aggregate state snapshots ([#121](https://github.com/commanded/commanded/pull/121)).
 - New `error/3` callback for process manager and deprecated `error/4` ([#124](https://github.com/commanded/commanded/pull/124))
 - Router support for identity prefix function.
+- Retry command execution on concurrency error ([#132](https://github.com/commanded/commanded/pull/132)).
 
 ## v0.15.1